refactor copy to clipboard button to make it more reusable
Usage example:

  <button data-button="copy-to-clipboard" data-content="foo">
    copy foo to clipboard
  </button>

This will copy "foo" to clipboard when user clicks the button.

It's safe to use other elements inside without adding data-content
to them as well. Event.currentTarget "always refers to the
element to which the event handler has been attached, as opposed to
Event.target, which identifies the element on which the event occurred
and which may be its direct descendent." (MDN)