fix: Split index route in public (index) and authenticated (start)
So far, `StartController::index()` rendered different content depending
on whether a user is authenticated or not. In the symfony security
firewall, the endpoint was marked as `PUBLIC_ACCESS`, which meant we
exposed content that requires authentication via a public endpoint.

This is now splitted:
* `index` remains the publically available index page
* `start` is the account overview page that requires authentication

Along the way, a few minor issues were fixed.